ui: remove gpu VM detail and Hdx customization
Hdx is no longer supported and has not been for a long time.
com.citrix.xenclient.xenmgr.vm.gpu no longer exist for that reason.

Note: This has nothing to do with PCI pass-through where a GPU on a
PCI/PCIe slot/bus is being passed through to the guest.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>